refactor: СhangeTeacherDialog переиспользован в Lesson/ContentTabStidents
